refactor: optimize thread argument structure and improve multithreaded analysis performance

+++ include/array_utils.h
@@ -5,7 +5,7 @@
 
 // Struct to hold thread segment statistics
 typedef struct SegmentStats {
-    int sum;
+    long long sum;
     int min;
     int max;
     int even_count;
